+++ packages/cli/CHANGELOG.md
@@ -4,8 +4,9 @@
 
 #### 🐞 Fixes
 
-- Fixed an issue where tasks depending on arbitrary project tasks, would fail to build a partial
+- Fixed an issue where tasks depending on arbitrary project tasks would fail to build a partial
   project graph.
+- Fixed an issue where task `deps` within global tasks would not apply exclude/rename overrides.
 - Fixed an issue with asset/binary files not working correctly with codegen templates.
 - Fixed an issue where `moon upgrade` would require a workspace.